requirements.txt
flask
PyJWT
jsonpickle
datetime
Flask-SQLAlchemy
mysqlclient
python-dotenv